feat(collections): stabilize iterable input for withoutAll
