+++ collections/without_all.ts
@@ -2,15 +2,19 @@
 // This module is browser compatible.
 
 /**
- * Returns an array excluding all given values.
+ * Returns an array excluding all given values from an iterable.
  *
- * @typeParam T The type of the array elements.
+ * @typeParam T The type of the elements in the iterable.
  *
- * @param array The array to exclude values from.
- * @param values The values to exclude from the array.
+ * @param iterable The iterable to exclude values from.
+ * @param values The values to exclude from the iterable.
  *
- * @returns A new array containing all elements from the given array except the
- * ones that are in the values array.
+ * @returns An array containing all elements from iterables except the
+ * ones that are in the values iterable.
+ *
+ * @remarks
+ * If both inputs are a {@linkcode Set}, and you want the difference as a
+ * {@linkcode Set}, you could use {@linkcode Set.prototype.difference} instead.
  *
  * @example Basic usage
  * ```ts
@@ -22,7 +26,14 @@
  * assertEquals(withoutList, [3]);
  * ```
  */
-export function withoutAll<T>(array: readonly T[], values: readonly T[]): T[] {
-  const toExclude = new Set(values);
-  return array.filter((it) => !toExclude.has(it));
+export function withoutAll<T>(iterable: Iterable<T>, values: Iterable<T>): T[] {
+  const excludedSet = new Set(values);
+  const result: T[] = [];
+  for (const value of iterable) {
+    if (excludedSet.has(value)) {
+      continue;
+    }
+    result.push(value);
+  }
+  return result;
 }
